gapp/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-cloud-tasks==1.5.0
googleapis-common-protos==1.51.0
sendgrid==6.3.1
google-cloud-storage
